refactor(ui): extract shared components into @kkhys/ui
spinner.astro (memo/lgtm) and budoux.astro + its cached parser
(me/trends) were byte-identical copies maintained in two apps each.
Move them into a source-consumed @kkhys/ui package so a fix lands once,
and re-point consumers; the parser ships as a plain-TS "./budoux"
export because .astro files can't be imported from .ts.
