+++ CLAUDE.md
@@ -18,6 +18,7 @@ Packages:
 - `packages/styles` — `@kkhys/styles`, uchu.css OKLCH palette + shared semantic tokens (`tokens.css`), base styles (`base.css`), and Satori hex mirrors (`colors`).
 - `packages/seo` — `@kkhys/seo`, BaseSEO / OpenGraph / TwitterCard Astro primitives.
 - `packages/og` — `@kkhys/og`, Satori OG image + favicon generators.
+- `packages/ui` — `@kkhys/ui`, shared Astro components (spinner, BudouX wrapper) + the cached BudouX parser.
 - `packages/analytics` — `@kkhys/analytics`, self-hosted Umami tracker component.
 - `packages/release` — `@kkhys/release`, date-based release tagging used by `scripts/release.ts`.
