+++ apps/lgtm/package.json
@@ -22,6 +22,7 @@
     "@kkhys/og": "workspace:*",
     "@kkhys/seo": "workspace:*",
     "@kkhys/styles": "workspace:*",
+    "@kkhys/ui": "workspace:*",
     "astro": "catalog:",
     "kiso.css": "catalog:",
     "sharp": "catalog:"
