+++ apps/me/CLAUDE.md
@@ -18,6 +18,7 @@ Consumed as source (no build step); this app supplies its own config via thin wr
 - `@kkhys/styles` — uchu.css OKLCH palette + shared semantic tokens / base styles; me keeps only prose- and code-specific tokens local
 - `@kkhys/seo` — BaseSEO / OpenGraph / TwitterCard primitives, wrapped in `src/components/seo/`
 - `@kkhys/og` — Satori OG image + favicon generators, wired in `src/pages/api/`
+- `@kkhys/ui` — shared Astro components and the helpers behind them; me uses the cached BudouX parser (`src/lib/rehype-budoux.ts`, OG image titles)
 
 ## Content System
