+++ apps/me/package.json
@@ -30,6 +30,7 @@
     "@kkhys/og": "workspace:*",
     "@kkhys/seo": "workspace:*",
     "@kkhys/styles": "workspace:*",
+    "@kkhys/ui": "workspace:*",
     "astro": "catalog:",
     "astro-expressive-code": "0.44.1",
     "kiso.css": "catalog:",
@@ -45,7 +46,6 @@
     "@types/unist": "3.0.3",
     "@vitest/coverage-v8": "catalog:",
     "bech32": "2.0.0",
-    "budoux": "catalog:",
     "fetch-site-metadata": "catalog:",
     "hast-util-from-html": "2.0.3",
     "hast-util-to-html": "9.0.5",
