+++ apps/me/src/features/blog/components/ui/blocks/memo-card.astro
@@ -1,7 +1,7 @@
 ---
 import { getImage, Image } from "astro:assets";
 import type { HTMLAttributes } from "astro/types";
-import Budoux from "#/components/budoux.astro";
+import Budoux from "@kkhys/ui/budoux.astro";
 import MemoIcon from "#/components/icons/memo.svg";
 import { extractMemoId, getMemoPost } from "#/lib/api/memo";
 import { blurLoadHandlers } from "#/utils/blur-load";
