+++ apps/me/src/features/blog/components/ui/entry-list.astro
@@ -1,6 +1,6 @@
 ---
 import type { HTMLAttributes } from "astro/types";
-import Budoux from "#/components/budoux.astro";
+import Budoux from "@kkhys/ui/budoux.astro";
 import MoveUpRightIcon from "#/components/icons/move-up-right.svg";
 import EmojiEyeCatch from "#/components/ui/emoji-eye-catch.astro";
 import SiteIcon from "#/features/blog/components/ui/site-icon.astro";
