+++ apps/me/src/features/blog/components/ui/post-navigation.astro
@@ -1,6 +1,6 @@
 ---
 import type { HTMLAttributes } from "astro/types";
-import Budoux from "#/components/budoux.astro";
+import Budoux from "@kkhys/ui/budoux.astro";
 import EmojiEyeCatch from "#/components/ui/emoji-eye-catch.astro";
 import type { PostNavItem } from "#/features/blog/utils/entry";
