+++ apps/me/src/pages/bucket-list.astro
@@ -1,6 +1,6 @@
 ---
 import { getEntry } from "astro:content";
-import Budoux from "#/components/budoux.astro";
+import Budoux from "@kkhys/ui/budoux.astro";
 import CheckIcon from "#/components/icons/check.svg";
 import SEO from "#/components/seo/seo.astro";
 import SiteBrand from "#/components/ui/site-brand.astro";
