+++ apps/me/src/pages/index.astro
@@ -1,5 +1,5 @@
 ---
-import Budoux from "#/components/budoux.astro";
+import Budoux from "@kkhys/ui/budoux.astro";
 import MoveUpRightIcon from "#/components/icons/move-up-right.svg";
 import JsonLd from "#/components/seo/json-ld.astro";
 import SEO from "#/components/seo/seo.astro";
