+++ apps/memo/package.json
@@ -21,6 +21,7 @@
     "@kkhys/og": "workspace:*",
     "@kkhys/seo": "workspace:*",
     "@kkhys/styles": "workspace:*",
+    "@kkhys/ui": "workspace:*",
     "astro": "catalog:",
     "kiso.css": "catalog:",
     "sharp": "catalog:"
