+++ apps/memo/src/components/infinite-scroll.astro
@@ -1,5 +1,5 @@
 ---
-import Spinner from "#/components/spinner.astro";
+import Spinner from "@kkhys/ui/spinner.astro";
 
 interface Props {
   containerId: string;
