--- apps/memo/src/components/spinner.astro
@@ -1,38 +1,0 @@
-<svg
-  xmlns="http://www.w3.org/2000/svg"
-  width={16}
-  height={16}
-  viewBox="0 0 24 24"
-  fill="none"
-  stroke="var(--c-sub)"
-  stroke-width="2"
-  stroke-linecap="round"
-  stroke-linejoin="round"
-  class="spinner"
-  role="status"
-  aria-label="Loading"
->
-  <path d="M12 2v4" />
-  <path d="m16.2 7.8 2.9-2.9" />
-  <path d="M18 12h4" />
-  <path d="m16.2 16.2 2.9 2.9" />
-  <path d="M12 18v4" />
-  <path d="m4.9 19.1 2.9-2.9" />
-  <path d="M2 12h4" />
-  <path d="m4.9 4.9 2.9 2.9" />
-</svg>
-
-<style>
-  .spinner {
-    animation: spin 1s linear infinite;
-  }
-
-  @keyframes spin {
-    from {
-      transform: rotate(0deg);
-    }
-    to {
-      transform: rotate(360deg);
-    }
-  }
-</style>
