+++ apps/trends/package.json
@@ -20,10 +20,9 @@
     "@kkhys/og": "workspace:*",
     "@kkhys/seo": "workspace:*",
     "@kkhys/styles": "workspace:*",
+    "@kkhys/ui": "workspace:*",
     "astro": "catalog:",
-    "budoux": "catalog:",
     "kiso.css": "catalog:",
-    "linkedom": "catalog:",
     "sharp": "catalog:"
   },
   "devDependencies": {
