--- apps/trends/src/components/budoux.astro
@@ -1,9 +1,0 @@
----
-import { budouxProcess } from "#/lib/budoux";
-
-const slotContent = await Astro.slots.render("default");
-
-const parsed = budouxProcess(slotContent);
----
-
-<Fragment set:html={parsed} />
