+++ apps/trends/src/components/digest-hero.astro
@@ -1,7 +1,7 @@
 ---
 import type { CollectionEntry } from "astro:content";
 
-import Budoux from "#/components/budoux.astro";
+import Budoux from "@kkhys/ui/budoux.astro";
 
 interface Props {
   digest: CollectionEntry<"runs">["data"]["digest"];
