+++ apps/trends/src/components/item-row.astro
@@ -1,7 +1,7 @@
 ---
 import type { CollectionEntry } from "astro:content";
 
-import Budoux from "#/components/budoux.astro";
+import Budoux from "@kkhys/ui/budoux.astro";
 import { formatHost } from "#/utils/host";
 import { splitParagraphs } from "#/utils/paragraphs";
 import { scoreLevel } from "#/utils/score";
