+++ apps/trends/src/pages/archive.astro
@@ -1,7 +1,7 @@
 ---
 import { getCollection } from "astro:content";
 
-import Budoux from "#/components/budoux.astro";
+import Budoux from "@kkhys/ui/budoux.astro";
 import SiteHeader from "#/components/site-header.astro";
 import BaseLayout from "#/layouts/base-layout.astro";
 import { formatRunDate, sortRunsByDateDesc } from "#/utils/runs";
