+++ packages/ui/src/budoux.astro
@@ -1,5 +1,5 @@
 ---
-import { budouxProcess } from "#/lib/budoux";
+import { budouxProcess } from "./budoux";
 
 const slotContent = await Astro.slots.render("default");
